+++ datapipe_app/__init__.py
@@ -7,20 +7,15 @@ from fastapi.staticfiles import StaticFiles
 from fastapi.middleware.cors import CORSMiddleware
 
 from datapipe.datatable import DataStore
-from datapipe.compute import build_compute, Catalog, Pipeline
+from datapipe.compute import Catalog, Pipeline, DatapipeApp
 
 import datapipe_app.api_v1alpha1 as api_v1alpha1
 
 
-class DatapipeApp(FastAPI):
+class DatapipeAPI(FastAPI, DatapipeApp):
     def __init__(self, ds: DataStore, catalog: Catalog, pipeline: Pipeline):
+        DatapipeApp.__init__(self, ds, catalog, pipeline)
         FastAPI.__init__(self)
-
-        self.ds = ds
-        self.catalog = catalog
-        self.pipeline = pipeline
-
-        self.steps = build_compute(ds, catalog, pipeline)
 
         self.add_middleware(
             CORSMiddleware,
@@ -32,11 +27,7 @@ class DatapipeApp(FastAPI):
 
         self.api = FastAPI()
 
-        self.api.mount(
-            "/v1alpha1",
-            api_v1alpha1.DatpipeAPIv1(ds, catalog, pipeline, self.steps),
-            name="v1alpha1"
-        )
+        self.api.mount("/v1alpha1", api_v1alpha1.DatpipeAPIv1(ds, catalog, pipeline, self.steps), name="v1alpha1")
 
         self.mount("/api", self.api, name="api")
         self.mount(